Refactor some internals of `ff`
